fix(jangar): split torghut evidence helpers
